Match drawShadow shadow-enable compare as signed int
CCameraPcs::m_fullScreenShadowEnabled is a u8 but the original compares
it in int domain (cmpwi, not cmplwi). Cast to int at the comparison site
to recover the signed compare, matching the target. drawShadow 99.12% -> 99.92%.

+++ src/p_chara.cpp
@@ -1307,7 +1307,7 @@ void CCharaPcs::drawMakeTexShadow()
  */
 void CCharaPcs::drawShadow()
 {
-    if (CameraPcs.m_fullScreenShadowEnabled == 0) {
+    if (static_cast<int>(CameraPcs.m_fullScreenShadowEnabled) == 0) {
         return;
     }
